Add identity precompile to the auction house
The identity precompile is an optimization to allow fast
copy from one area of memory to another.  For some reason,
Etherscan verification seems to assume this optimization somewhere
when it compiles the contract so NOT including this means
verification will bork!

This fix was checked against the online compiler, which seems to work
fine with the NFT contract.  Still need to check with an actual
deployment.